Create LICENSE
tl;dr -- providing a missing license file, copyright information is best-guess, go ahead and modify the PR with accurate information

The `LICENSE` file appears to be missing or was accidentally removed (I didn't dig around to much to see what happened.) Since the `README` declares MIT license and points to the `LICENSE` file in the source root, I figure this MIT file with the copyright dates covering the length of the codebase would be the right thing, yes?

Regarding copyright ownership mentioned there --that practice of naming an entity and others ("et al") is one I have seen and used before to simplify the list. This is because the commit history holds the details about who created what. (A variation is "Copyright (C) 2022-2025 Kit Project contributors", and the contributors are by definition those in the commit history.)

The reason for picking Alex's name is this project shows the signs of being a personal project and not on behalf of an employer. Where open source is written as works-for-hire, the practice of putting the employer in as the copyright holder is commonly used, and the licensing by that copyright holder provides the individual contributors the assurance of using and contributing to this code base going forward (even if they move on to a different employer, etc.)

Signed-off-by: Karsten Wade <quaid@iquaid.org>