feat: Add support for default task
This change adds the ability to specify a default task in the `tasks.yaml` file. If no task is specified on the command line, the default task will be run.

This is implemented by adding a `Default` field to the `Task` struct in `internal/types/task.go`. The `main.go` file is updated to check for a default task if no task is specified on the command line.

The documentation and schema are also updated to reflect the new `default` field.